Migrate Base64 away from Msebera (in Android), Fixes AB#3077784 (#2389)
[AB#3077784](https://identitydivision.visualstudio.com/fac9d424-53d2-45c0-91b5-ef6ba7a6bf26/_workitems/edit/3077784)
(Part 2)

## Background

Our entire code base is written to be compatible with
Android.util.Base64, which is a fork of a legacy version of apache
commons-codec.

This means Android OS ships with apache commons-codec, and we cannot
have that as a dependency in common4j project (otherwise the classloader
would throw an error).

The safest, previous workaround is to consume [msebera
HttpClient](https://github.com/smarek/httpclient-android), which also
contains commons-codec, but under a different namespace. [(This is
officially suggested by Apache)
](https://hc.apache.org/httpcomponents-client-4.5.x/android.html)

## Issue 

msebera library is no longer maintained, and it was recently flagged
with an apk tool.

Even if the flagged part is NOT the code path we use. It's better that
we migrate away from it (so that our customer don't see the false alarm)

There are multiple ways to tackle this
1. (Least risky) Make the android project consume Android.util.base64,
keep msebera for Linux (which is already in maintainance mode) and test
classes
2. Migrate to other libraries. (e.g. guava)
   - NOTE: Java's Base64 is not an option because it requires min API 26

I decide to go with 1 because 
- the other option might increase the final apk/aar size by a lot. 
- It would also introduce a breaking change (e.g. The value in storage -
before upgrade - was encoded with msebera, and might not be compatible
with the new library.
- The new library might also handle error cases/unexpected format
differently.

## Changes

All the Base64 usage is now consolidated into the Base64Util class.
(most of the changes are under this bucket).

This Base64Util class will use a classloader to pick the implementation
to use.
- In Android modules, it would load `AndroidBase64`
- In Linux Broker, it would load `MseberaBase64ForLinux`
- In Common4j's own unit test, it would load
`MseberaBase64ForCommon4jTests` (from common4j's testImplementation)
- In Broker4j's own unit test, it would load
`MseberaBase64ForBroker4jTests` (from common4j's testImplementation)
- In other tests, it would load `MseberaBase64ForTestUtils` (from
:testutils)