Make userHandle response field optional (#2560)
### Summary
The Windows app team found a bug while testing cross-device passkey
auth: https://microsoft.visualstudio.com/OS/_workitems/edit/55529057

In the logs, I see that the first attempt fails due to the userHandle
attribute being missing in the response from CredMan. We currently have
this attribute set as required, since the server side had mentioned that
this attribute was required for them, but looking at the official
WebAuthN spec, userHandle is only required if allowCredentials is NOT
provided. If it is (like in the first attempt) then userHandle is
optional:
https://w3c.github.io/webauthn/#iface-authenticatorassertionresponse

I've confirmed with Authenticator that they always do send userHandle,
and our current thinking is that some middle layer (since this is
cross-device auth) might be removing that value from the final response.
I'm not sure why this might be the case, but this is perfectly fine
according to the WebAuthN spec. Therefore, we will follow the spec as
well and will not block the response on userHandle.