Managed profile Android util method (#2561)
### Summary
This PR creates an `AndroidPlatformUtil` method named `isInWorkProfile`,
which returns a boolean based on if the host app is in a work profile.
For Android level 30 (Android 11/R) and above, we are able to use the
simple `isManagedProfile` method of `UserManager`. For Android level 21
(Android 5/Lollipop) and above, we get the `DevicePolicyManager`
instance and get a list of the active admins. If any of these active
admins are found to be the profile owner app, then we know that the
calling app is in a work profile. (We do something similar in
`InstallCertActivity` of the WPJ feature). If the device is below
Android level 21 (should be very rare), we always return false.
I tested this method out using testDPC and made sure both paths of the
logic return the correct values.