Sign In With Google for MSA federation (No broker), Fixes AB#3100601 (#2551)
This pull request adds the Sign In With Google flow to support MSA
federation in Android Common

1. Add abstract class interface to support federated sign ins
(FederatedSignInProviderFactory, IFederatedSignInProvider,
FederatedSignInParameters, FederatedCredential). This can be used for
specific sign in providers like Google. These are thin classes.
2. Add implementation for google (GoogleSignInProvider,
SignInWithGoogleParameters, SignInWithGoogleCredential).
GoogleSignInProvider uses credential manager to get allow user to sign
select google account or added new google account on device and get
credentials for it. Credential is essentially google idtoken. This is
abstracted in SignInWithGoogleCredential and returned to the caller.
SignInWithGoogleParameters include inputs that are used to run the sign
in flow, contains activity/context and whether to use bottom sheet UX or
not. By default dialog is launched (based on Sign in with google button
which is part of UX/PM spec).
3. SignInWithGoogleApi is entry point to start sign in flow and utilizes
above internal classes. This would be exposed to OneAuth.
4. OneAuth interacts with common for performing authorization using
WebView. To launch authorization UX it uses,
AuthorizationActivityFactory methods to get an intent which is later on
launched in the activity or fragment. We expose two methods in this
class to allow sign in with google.

AuthorizationActivityFactory.signInWithGoogleAndGetAuthorizationActivityIntent
- first uses google sign in classes to collect google credentials and
prepares intent based on that. This is intent is returned to caller.
AuthorizationActivityFactory.getAuthorizationActivityIntent(...,
SignInWithGoogleCredential) - If the caller has acquired google sign in
credential previously using SignInWithGoogleApi, then it should call
this method to create intent which would use the credential for running
authorization flow.
The authorization flow is scoped to MSA (v2/consumers endpoint)
5. Broker should be able to use the code in common to get the
credential.
6. Added unit tests cases.



![image](https://github.com/user-attachments/assets/004eeb84-6235-4d3a-be22-36ab7b58e0b4)

Fixes
[AB#3100601](https://identitydivision.visualstudio.com/fac9d424-53d2-45c0-91b5-ef6ba7a6bf26/_workitems/edit/3100601)